pyproject.toml
[project]
name = "mcengine"
version = "0.1.0"
description = "Monte Carlo Engine"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "rq>=2.6.1",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
